+++ backend/src/baserow/config/settings/base.py
@@ -22,6 +22,7 @@ INSTALLED_APPS = [
 
     'rest_framework',
     'corsheaders',
+    'mjml',
 
     'baserow.core',
     'baserow.api.v0',
@@ -143,3 +144,17 @@ JWT_AUTH = {
 }
 
 DATABASE_ROUTERS = ('baserow.contrib.database.database_routers.TablesDatabaseRouter',)
+
+MJML_BACKEND_MODE = 'tcpserver'
+MJML_TCPSERVERS = [
+    ('mjml', 28101),
+]
+
+PUBLIC_BACKEND_DOMAIN = 'localhost:8000'
+PUBLIC_BACKEND_URL = 'http://localhost:8000'
+PUBLIC_WEB_FRONTEND_DOMAIN = 'localhost:3000'
+PUBLIC_WEB_FRONTEND_URL = 'http://localhost:3000'
+
+FROM_EMAIL = 'no-reply@localhost'
+
+RESET_PASSWORD_TOKEN_MAX_AGE = 60 * 60 * 48  # 48 hours